Remove ESSearchHit sourceAsMap deprecated field
We don't use it anymore now we have JsonPath.
