+++ integration-tests/src/test/java/fr/pilato/elasticsearch/crawler/fs/test/integration/elasticsearch/FsCrawlerRestIT.java
@@ -19,47 +19,28 @@
 
 package fr.pilato.elasticsearch.crawler.fs.test.integration.elasticsearch;
 
+import com.jayway.jsonpath.JsonPath;
+import com.jayway.jsonpath.PathNotFoundException;
 import fr.pilato.elasticsearch.crawler.fs.client.ESSearchHit;
 import fr.pilato.elasticsearch.crawler.fs.client.ESSearchRequest;
 import fr.pilato.elasticsearch.crawler.fs.client.ESSearchResponse;
-import fr.pilato.elasticsearch.crawler.fs.client.ESTermQuery;
-import fr.pilato.elasticsearch.crawler.fs.client.ElasticsearchClientException;
 import fr.pilato.elasticsearch.crawler.fs.framework.TimeValue;
 import fr.pilato.elasticsearch.crawler.fs.framework.Version;
 import fr.pilato.elasticsearch.crawler.fs.rest.DeleteResponse;
-import fr.pilato.elasticsearch.crawler.fs.rest.RestServer;
 import fr.pilato.elasticsearch.crawler.fs.rest.ServerStatusResponse;
 import fr.pilato.elasticsearch.crawler.fs.rest.UploadResponse;
-import fr.pilato.elasticsearch.crawler.fs.service.FsCrawlerDocumentService;
-import fr.pilato.elasticsearch.crawler.fs.service.FsCrawlerDocumentServiceElasticsearchImpl;
-import fr.pilato.elasticsearch.crawler.fs.service.FsCrawlerDocumentServiceWorkplaceSearchImpl;
-import fr.pilato.elasticsearch.crawler.fs.service.FsCrawlerManagementServiceElasticsearchImpl;
-import fr.pilato.elasticsearch.crawler.fs.settings.FsCrawlerValidator;
 import fr.pilato.elasticsearch.crawler.fs.settings.FsSettings;
 import fr.pilato.elasticsearch.crawler.fs.settings.Rest;
 import fr.pilato.elasticsearch.crawler.fs.test.integration.AbstractRestITCase;
-import jakarta.ws.rs.client.WebTarget;
-import jakarta.ws.rs.core.MediaType;
-import org.glassfish.jersey.media.multipart.FormDataMultiPart;
-import org.glassfish.jersey.media.multipart.file.FileDataBodyPart;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Predicate;
 
-import static com.carrotsearch.randomizedtesting.RandomizedTest.rarely;
-import static fr.pilato.elasticsearch.crawler.fs.framework.FsCrawlerUtil.INDEX_SUFFIX_FOLDER;
-import static fr.pilato.elasticsearch.crawler.fs.framework.FsCrawlerUtil.copyDirs;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
@@ -67,72 +48,11 @@ import static org.hamcrest.Matchers.*;
 @SuppressWarnings("ALL")
 public class FsCrawlerRestIT extends AbstractRestITCase {
 
-    private Path currentTestTagDir;
-    private FsCrawlerManagementServiceElasticsearchImpl managementService;
-    private FsCrawlerDocumentService documentService;
-
-    @Before
-    public void copyTags() throws IOException {
-        Path testResourceTarget = rootTmpDir.resolve("resources");
-        if (Files.notExists(testResourceTarget)) {
-            Files.createDirectory(testResourceTarget);
-        }
-
-        String currentTestName = getCurrentTestName();
-        // We copy files from the src dir to the temp dir
-        String url = getUrl("tags", currentTestName);
-        Path from = Paths.get(url);
-
-        currentTestTagDir = testResourceTarget.resolve(currentTestName + ".tags");
-        if (Files.exists(from)) {
-            staticLogger.debug("  --> Copying test resources from [{}]", from);
-            copyDirs(from, currentTestTagDir);
-            staticLogger.debug("  --> Tags ready in [{}]", currentTestTagDir);
-        }
-    }
-
-    @Before
-    public void startRestServer() throws Exception {
-        FsSettings fsSettings = FsSettings.builder(getCrawlerName())
+    public FsSettings getFsSettings() {
+        return FsSettings.builder(getCrawlerName())
                 .setRest(new Rest("http://127.0.0.1:" + testRestPort + "/fscrawler"))
                 .setElasticsearch(elasticsearchWithSecurity)
                 .build();
-        fsSettings.getElasticsearch().setIndex(getCrawlerName());
-        FsCrawlerValidator.validateSettings(logger, fsSettings, true);
-
-        this.managementService = new FsCrawlerManagementServiceElasticsearchImpl(metadataDir, fsSettings);
-
-        if (fsSettings.getWorkplaceSearch() == null) {
-            // The documentService is using the esSearch instance
-            this.documentService = new FsCrawlerDocumentServiceElasticsearchImpl(metadataDir, fsSettings);
-        } else {
-            // The documentService is using the wpSearch instance
-            this.documentService = new FsCrawlerDocumentServiceWorkplaceSearchImpl(metadataDir, fsSettings);
-        }
-
-        managementService.start();
-        documentService.start();
-
-        RestServer.start(fsSettings, managementService, documentService);
-
-        logger.info(" -> Removing existing index [{}]", getCrawlerName() + "*");
-        managementService.getClient().deleteIndex(getCrawlerName());
-        managementService.getClient().deleteIndex(getCrawlerName() + INDEX_SUFFIX_FOLDER);
-
-        logger.info(" -> Creating index [{}]", fsSettings.getElasticsearch().getIndex());
-    }
-
-    @After
-    public void stopRestServer() throws IOException {
-        RestServer.close();
-        if (managementService != null) {
-            managementService.close();
-            managementService = null;
-        }
-        if (documentService != null) {
-            documentService.close();
-            documentService = null;
-        }
     }
 
     @Test
@@ -160,8 +80,7 @@ public class FsCrawlerRestIT extends AbstractRestITCase {
         ESSearchResponse response = countTestHelper(new ESSearchRequest().withIndex(getCrawlerName()), Files.list(from).count(), null, TimeValue
                 .timeValueMinutes(2));
         for (ESSearchHit hit : response.getHits()) {
-            assertThat(hit.getSourceAsMap(), hasKey("file"));
-            assertThat((Map<String, Object>) hit.getSourceAsMap().get("file"), hasKey("extension"));
+            assertThat(JsonPath.read(hit.getSource(), "$.file.extension"), notNullValue());
         }
     }
 
@@ -192,8 +111,7 @@ public class FsCrawlerRestIT extends AbstractRestITCase {
         ESSearchResponse response = countTestHelper(new ESSearchRequest().withIndex(getCrawlerName()), number.longValue(), null, TimeValue
                 .timeValueMinutes(2));
         for (ESSearchHit hit : response.getHits()) {
-            assertThat(hit.getSourceAsMap(), hasKey("file"));
-            assertThat((Map<String, Object>) hit.getSourceAsMap().get("file"), hasKey("extension"));
+            assertThat(JsonPath.read(hit.getSource(), "$.file.extension"), notNullValue());
         }
     }
 
@@ -262,8 +180,7 @@ public class FsCrawlerRestIT extends AbstractRestITCase {
         ESSearchResponse response = countTestHelper(new ESSearchRequest().withIndex(getCrawlerName()), number.longValue(), null, TimeValue
                 .timeValueMinutes(2));
         for (ESSearchHit hit : response.getHits()) {
-            assertThat(hit.getSourceAsMap(), hasKey("file"));
-            assertThat((Map<String, Object>) hit.getSourceAsMap().get("file"), hasKey("extension"));
+            assertThat(JsonPath.read(hit.getSource(), "$.file.extension"), notNullValue());
         }
 
         // We can now remove all docs
@@ -299,8 +216,7 @@ public class FsCrawlerRestIT extends AbstractRestITCase {
         ESSearchResponse response = countTestHelper(new ESSearchRequest().withIndex(index), Files.list(from).count(), null, TimeValue
                 .timeValueMinutes(2));
         for (ESSearchHit hit : response.getHits()) {
-            assertThat(hit.getSourceAsMap(), hasKey("file"));
-            assertThat((Map<String, Object>) hit.getSourceAsMap().get("file"), hasKey("extension"));
+            assertThat(JsonPath.read(hit.getSource(), "$.file.extension"), notNullValue());
         }
     }
 
@@ -323,250 +239,45 @@ public class FsCrawlerRestIT extends AbstractRestITCase {
 
         // Let's test every single document that has been enriched
         checkDocument("add_external.txt", hit -> {
-            String content = (String) hit.getSourceAsMap().get("content");
-            assertThat(content, containsString("This file content will be extracted"));
-
-            assertThat(hit.getSourceAsMap(), hasKey("file"));
-            assertThat((Map<String, Object>) hit.getSourceAsMap().get("file"), hasKey("extension"));
-
-            assertThat(hit.getSourceAsMap(), not(hasKey("meta")));
-
-            assertThat(hit.getSourceAsMap(), hasKey("external"));
-            Map<String, Object> external = (Map<String, Object>) hit.getSourceAsMap().get("external");
-            assertThat(external, hasKey("tenantId"));
-            Integer tenant = (Integer) external.get("tenantId");
-            assertThat(tenant, is(23));
-
-            assertThat(external, hasKey("company"));
-            String company = (String) external.get("company");
-            assertThat(company, is("shoe company"));
-
-            assertThat(external, hasKey("daysOpen"));
-            List<String> daysOpen = (List<String>) external.get("daysOpen");
-            assertThat(daysOpen.size(), is(5));
-            assertThat(daysOpen.get(0), is("Mon"));
-            assertThat(daysOpen.get(4), is("Fri"));
-
-            assertThat(external, hasKey("products"));
-            List<Object> products = (List<Object>) external.get("products");
-            assertThat(products.size(), is(2));
-
-            Map<String, Object> nike = (Map<String, Object>) products.get(0);
-            assertThat(nike, hasKey("brand"));
-            assertThat(nike, hasKey("size"));
-            assertThat(nike, hasKey("sub"));
-            String brand = (String) nike.get("brand");
-            Integer size = (Integer) nike.get("size");
-            String sub = (String) nike.get("sub");
-            assertThat(brand, is("nike"));
-            assertThat(size, is(41));
-            assertThat(sub, is("Air MAX"));
+            assertThat(JsonPath.read(hit.getSource(), "$.content"), containsString("This file content will be extracted"));
+            assertThat(JsonPath.read(hit.getSource(), "$.file.extension"), notNullValue());
+            expectThrows(PathNotFoundException.class, () -> JsonPath.read(hit.getSource(), "$.meta"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.tenantId"), is(23));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.company"), is("shoe company"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.daysOpen[0]"), is("Mon"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.daysOpen[4]"), is("Fri"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[0].brand"), is("nike"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[0].size"), is(41));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[0].sub"), is("Air MAX"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[1].brand"), is("reebok"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[1].size"), is(43));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[1].sub"), is("Pump"));
         });
         checkDocument("replace_content_and_external.txt", hit -> {
-            String content = (String) hit.getSourceAsMap().get("content");
-            assertThat(content, is("OVERWRITTEN CONTENT"));
-
-            assertThat(hit.getSourceAsMap(), hasKey("file"));
-            assertThat((Map<String, Object>) hit.getSourceAsMap().get("file"), hasKey("extension"));
-
-            assertThat(hit.getSourceAsMap(), not(hasKey("meta")));
-
-            assertThat(hit.getSourceAsMap(), hasKey("external"));
-            Map<String, Object> external = (Map<String, Object>) hit.getSourceAsMap().get("external");
-            assertThat(external, hasKey("tenantId"));
-            Integer tenant = (Integer) external.get("tenantId");
-            assertThat(tenant, is(23));
-
-            assertThat(external, hasKey("company"));
-            String company = (String) external.get("company");
-            assertThat(company, is("shoe company"));
-
-            assertThat(external, hasKey("daysOpen"));
-            List<String> daysOpen = (List<String>) external.get("daysOpen");
-            assertThat(daysOpen.size(), is(5));
-            assertThat(daysOpen.get(0), is("Mon"));
-            assertThat(daysOpen.get(4), is("Fri"));
-
-            assertThat(external, hasKey("products"));
-            List<Object> products = (List<Object>) external.get("products");
-            assertThat(products.size(), is(2));
-
-            Map<String, Object> nike = (Map<String, Object>) products.get(0);
-            assertThat(nike, hasKey("brand"));
-            assertThat(nike, hasKey("size"));
-            assertThat(nike, hasKey("sub"));
-            String brand = (String) nike.get("brand");
-            Integer size = (Integer) nike.get("size");
-            String sub = (String) nike.get("sub");
-            assertThat(brand, is("nike"));
-            assertThat(size, is(41));
-            assertThat(sub, is("Air MAX"));
+            assertThat(JsonPath.read(hit.getSource(), "$.content"), is("OVERWRITTEN CONTENT"));
+            assertThat(JsonPath.read(hit.getSource(), "$.file.extension"), notNullValue());
+            expectThrows(PathNotFoundException.class, () -> JsonPath.read(hit.getSource(), "$.meta"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.tenantId"), is(23));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.company"), is("shoe company"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.daysOpen[0]"), is("Mon"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.daysOpen[4]"), is("Fri"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[0].brand"), is("nike"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[0].size"), is(41));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[0].sub"), is("Air MAX"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[1].brand"), is("reebok"));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[1].size"), is(43));
+            assertThat(JsonPath.read(hit.getSource(), "$.external.products[1].sub"), is("Pump"));
         });
         checkDocument("replace_content_only.txt", hit -> {
-            String content = (String) hit.getSourceAsMap().get("content");
-            assertThat(content, is("OVERWRITTEN CONTENT"));
-
-            assertThat(hit.getSourceAsMap(), hasKey("file"));
-            assertThat((Map<String, Object>) hit.getSourceAsMap().get("file"), hasKey("extension"));
-
-            assertThat(hit.getSourceAsMap(), not(hasKey("meta")));
-
-            assertThat(hit.getSourceAsMap(), not(hasKey("external")));
+            assertThat(JsonPath.read(hit.getSource(), "$.content"), is("OVERWRITTEN CONTENT"));
+            assertThat(JsonPath.read(hit.getSource(), "$.file.extension"), notNullValue());
+            expectThrows(PathNotFoundException.class, () -> JsonPath.read(hit.getSource(), "$.meta"));
+            expectThrows(PathNotFoundException.class, () -> JsonPath.read(hit.getSource(), "$.external"));
         });
         checkDocument("replace_meta_only.txt", hit -> {
-            String content = (String) hit.getSourceAsMap().get("content");
-            assertThat(content, containsString("This file content will be extracted"));
-
-            assertThat(hit.getSourceAsMap(), hasKey("meta"));
-            Map<String, Object> meta = (Map<String, Object>) hit.getSourceAsMap().get("meta");
-            assertThat(meta, hasKey("raw"));
-            Map<String, Object> raw = (Map<String, Object>) meta.get("raw");
-            assertThat(raw, hasKey("resourceName"));
-            assertThat(raw.get("resourceName"), is("another-file-name.txt"));
-
-            assertThat(hit.getSourceAsMap(), not(hasKey("external")));
+            assertThat(JsonPath.read(hit.getSource(), "$.content"), containsString("This file content will be extracted"));
+            assertThat(JsonPath.read(hit.getSource(), "$.meta.raw.resourceName"), is("another-file-name.txt"));
+            expectThrows(PathNotFoundException.class, () -> JsonPath.read(hit.getSource(), "$.external"));
         });
-    }
-
-    private void checkDocument(String filename, HitChecker checker) throws IOException, ElasticsearchClientException {
-        ESSearchResponse response = documentService.search(new ESSearchRequest()
-                .withIndex(getCrawlerName())
-                .withESQuery(new ESTermQuery("file.filename", filename)));
-
-        assertThat(response.getTotalHits(), is(1L));
-        ESSearchHit hit = response.getHits().get(0);
-        logger.debug("For [file.filename:{}], we got: {}", filename, hit.getSource());
-
-        checker.check(hit);
-    }
-
-    private interface HitChecker {
-        void check(ESSearchHit hit);
-    }
-
-    private static final Map<String, Object> debugOption = new HashMap<>();
-
-    static {
-        debugOption.put("debug", true);
-        debugOption.put("simulate", true);
-    }
-
-    public static UploadResponse uploadFile(WebTarget target, Path file) {
-        return uploadFileUsingApi(target, file, null, null, null, null);
-    }
-
-    public static UploadResponse uploadFileOnIndex(WebTarget target, Path file, String index) {
-        return uploadFileUsingApi(target, file, null, index, null, null);
-    }
-
-    public static UploadResponse uploadFileWithId(WebTarget target, Path file, String id) {
-        return uploadFileUsingApi(target, file, null, null, null, id);
-    }
-
-    public static UploadResponse uploadFile(WebTarget target, Path file, Path tagsFile, String index) {
-        return uploadFileUsingApi(target, file, tagsFile, index, null, null);
-    }
-
-    public static UploadResponse uploadFileUsingApi(WebTarget target, Path file, Path tagsFile, String index, String api, String id) {
-        assertThat(Files.exists(file), is(true));
-
-        Map<String, Object> params = new HashMap<>();
-
-        if (api == null) {
-            api = "/_document";
-        }
-
-        // MediaType of the body part will be derived from the file.
-        FileDataBodyPart filePart = new FileDataBodyPart("file", file.toFile(), MediaType.APPLICATION_OCTET_STREAM_TYPE);
-        FormDataMultiPart mp = new FormDataMultiPart();
-        mp.bodyPart(filePart);
-
-        if (index != null) {
-            mp.field("index", index);
-            // Sadly this does not work
-            /*
-            if (rarely()) {
-                staticLogger.info("Force index name to {} using a form field", index);
-                mp.field("index", index);
-            } else {
-                staticLogger.info("Force index name to {} using a query string parameter", index);
-                params.put("index", index);
-            }
-            */
-        }
-
-        if (id != null) {
-            mp.field("id", id);
-            // Sadly this does not work
-            /*
-            if (rarely()) {
-                staticLogger.info("Force id to {} using a form field", id);
-                mp.field("id", id);
-            } else {
-                staticLogger.info("Force id to {} using a query string parameter", id);
-                params.put("id", id);
-            }
-             */
-        }
-
-        if (tagsFile != null) {
-            FileDataBodyPart tagsFilePart = new FileDataBodyPart("tags", tagsFile.toFile(), MediaType.APPLICATION_OCTET_STREAM_TYPE);
-            mp.bodyPart(tagsFilePart);
-        }
-
-        if (staticLogger.isDebugEnabled()) {
-            staticLogger.debug("Rest response: {}", post(target, api, mp, String.class, debugOption));
-        }
-
-        return post(target, api, mp, UploadResponse.class, params);
-    }
-
-    public static UploadResponse putDocument(WebTarget target, Path file, Path tagsFile, String index, String id) {
-        assertThat(Files.exists(file), is(true));
-
-        Map<String, Object> params = new HashMap<>();
-
-        // MediaType of the body part will be derived from the file.
-        FileDataBodyPart filePart = new FileDataBodyPart("file", file.toFile(), MediaType.APPLICATION_OCTET_STREAM_TYPE);
-        FormDataMultiPart mp = new FormDataMultiPart();
-        mp.bodyPart(filePart);
-
-        if (index != null) {
-            if (rarely()) {
-                staticLogger.info("Force index name to {} using a form field", index);
-                mp.field("index", index);
-            } else {
-                staticLogger.info("Force index name to {} using a query string parameter", index);
-                params.put("index", index);
-            }
-        }
-
-        if (tagsFile != null) {
-            FileDataBodyPart tagsFilePart = new FileDataBodyPart("tags", tagsFile.toFile(), MediaType.APPLICATION_OCTET_STREAM_TYPE);
-            mp.bodyPart(tagsFilePart);
-        }
-
-        return put(target, "/_document/" + id, mp, UploadResponse.class, params);
-    }
-
-    public static DeleteResponse deleteDocument(WebTarget target, String index, String id, String filename, String api) {
-        if (id != null) {
-            api = api + "/" + id;
-            staticLogger.info("Using id {}. Api is now {}", id, api);
-        }
-
-        Map<String, Object> options = new HashMap<>();
-
-        if (index != null) {
-            staticLogger.info("Using index {}", index);
-            options.put("index", index);
-        }
-
-        if (filename != null) {
-            staticLogger.info("Using filename {}", filename);
-            options.put("filename", filename);
-        }
-
-        return delete(target, api, DeleteResponse.class, options);
     }
 }
